mk-docker.nix: use `streamLayeredImage` instead of `buildImage`
Doesn't require assembling the image on the Nix side before passing to
Docker, saving disk and compute, due to streaming.
Due to layering, we should also be able to take advantage of layer caching on the Docker side, saving additional rebuild time.
(This would also make it more feasible to include the code in the image
rather than on disk as we currently do for ICH and ASIST-TBI.)

Removed disk and VM memory size options as streamLayeredImage doesn't
support these.